requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
airportsdata
altair
annotated-types
anyio
arxiv
asgiref
astor
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
blake3
blinker
bs4
build
cachetools
certifi
chardet
charset-normalizer
chroma-hnswlib
chromadb
click
cloudpickle
coloredlogs
compressed-tensors
cupy-cuda12x
dataclasses-json
depyf
dill
diskcache
distro
dnspython
duckduckgo_search
durationpy
einops
email_validator
exceptiongroup
fastapi
fastapi-cli
fastrlock
feedparser
filelock
flatbuffers
frozenlist
fsspec
gguf
gitdb
GitPython
google-auth
googleapis-common-protos
greenlet
grpcio
h11
hf-xet
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
interegular
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-ollama
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
lark
llguidance
llvmlite
lm-format-enforcer
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
mistral_common
mmh3
mpmath
msgpack
msgspec
multidict
mypy_extensions
narwhals
nest-asyncio
networkx
ninja
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
ollama
onnxruntime
openai
opencv-python-headless
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-semantic-conventions-ai
opentelemetry-util-http
orjson
ormsgpack
outlines
outlines_core
overrides
packaging
pandas
partial-json-parser
pillow
posthog
primp
prometheus-fastapi-instrumentator
prometheus_client
propcache
protobuf
psutil
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pycountry
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-json-logger
python-multipart
pytz
PyYAML
pyzmq
ray
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rpds-py
rsa
safetensors
scipy
sentencepiece
sgmllib3k
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
streamlit
streamlit-flow-component
sympy
tenacity
tiktoken
tokenizers
toml
tomli
torch
torchaudio
torchvision
tornado
tqdm
transformers
triton
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
uvloop
vllm
watchdog
watchfiles
websocket-client
websockets
wikipedia
wrapt
xformers
xgrammar
xinference-client
yarl
zipp